Edits in run-line formulas
According to Knuth's The TeX Book, one should not place ternary
operators (like + or -) at the end of a line in formulas that span
multiple lines. Rather, one should put it at the start so that it reads
easier. The reader should immediately be able to see whether the line
s/he is reading needs to be added or subtracted.